Allows transaction name to be changed after start
The name of the transaction won't always be known when start is called.
In a raw plug (non-Phoenix), the name of the transaction may not be known
until register_before_send's callback is executed. This changes allows
the transaction name to be updated/set during that callback before the
transaction is finished.