Allows running when not configured
Instead of raising an exception, this change lets the library run
and the agent will not push if it isn't configured. This allows the
library to be used in many different environments including those
that shouldn't be writing to New Relic (with minimal to no config
changes required from the consuming application).

This makes using this library similar to using logging. The app
always does the same thing (e.g. logging), but the library decides
when to publish or not. In this example, by when New Relic is configured
versus the logging level for logging.